+++ backend/src/main/java/com/example/demo/TicketCategoryRepository.java
@@ -2,6 +2,6 @@ package com.example.demo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketCategoryRepository extends JpaRepository<TicketCategory, Integer> {
+public interface TicketCategoryRepository extends JpaRepository<TicketCategory, Long> {
     
 }
